% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.cont.table.data.R
\name{gen.cont.table.data}
\alias{gen.cont.table.data}
\title{gen.cont.table.data}
\usage{
gen.cont.table.data(n, A, B, tbl = FALSE, rho)
}
\arguments{
\item{n}{sample size}

\item{A}{vector of values of first categorical variable}

\item{B}{vector of values of second categorical variable}

\item{tbl}{should output be a table}

\item{rho}{correlation between A and B}
}
\value{
A matrix with two columns
}
\description{
This function generates data for problems that use contingency tables
}
\examples{
gen.cont.table.data(10, c("a", "b"), 1:3, rho=0.9)
}
