% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genquiz.R
\name{genquiz}
\alias{genquiz}
\title{genquiz}
\usage{
genquiz(B = 1, fun, folder, problem = 0, Show = FALSE, ...)
}
\arguments{
\item{B}{how many quizzes?}

\item{fun}{name of the R routine that makes a quiz}

\item{folder}{where is the .R located?}

\item{problem}{(optional) which problem should be done?}

\item{Show}{(optional)want to see what it looks like?}

\item{...}{further arguments passed to fun}
}
\value{
None
}
\description{
This function generates an xml file for import into moodle
}
\examples{
# not run
# genquiz(10, example1)
}
