% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle.table.R
\name{moodle.table}
\alias{moodle.table}
\title{moodle.table}
\usage{
moodle.table(x, DoRowNames = FALSE, DoBorder = FALSE, ncols = 10)
}
\arguments{
\item{x}{df or vector}

\item{DoRowNames}{print row names?}

\item{DoBorder}{print border?}

\item{ncols}{for vectors, how many items per row?}
}
\value{
A character vector with html code
}
\description{
This function takes a data frame or vector and generate the html code to display it in a
moodle quiz
}
\examples{
moodle.table(round(rnorm(50), 1))
moodle.table(mtcars)
}
