% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample10.R
\docType{data}
\name{moodleRexample10}
\alias{moodleRexample10}
\title{Info for moodlequizR example 10}
\format{
A list
\describe{
  \item{quizname}{example10}
  \item{category}{MoodlequizR Examples / 10}
  ...
}
}
\usage{
moodleRexample10
}
\description{
A dataset containing the info to create the xml file for example 10
}
\keyword{datasets}
