% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample11.R
\docType{data}
\name{moodleRexample11}
\alias{moodleRexample11}
\title{Info for moodleR example 11}
\format{
A list
\describe{
  \item{quizname}{example11}
  \item{category}{MoodleR Examples / 11}
  ...
}
}
\usage{
moodleRexample11
}
\description{
A dataset containing the info to create the xml file for example 11
}
\keyword{datasets}
