% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample12.R
\docType{data}
\name{moodleRexample12}
\alias{moodleRexample12}
\title{Info for moodleR example 12}
\format{
A list
\describe{
  \item{quizname}{example12}
  \item{category}{MoodleR Examples / 12}
  ...
}
}
\usage{
moodleRexample12
}
\description{
A dataset containing the info to create the xml file for example 12
}
\keyword{datasets}
