% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample13.R
\docType{data}
\name{moodleRexample13}
\alias{moodleRexample13}
\title{Info for moodlequizR example 13}
\format{
A list
\describe{
  \item{quizname}{example13}
  \item{category}{MoodlequizR Examples / 13}
  ...
}
}
\usage{
moodleRexample13
}
\description{
A dataset containing the info to create the xml file for example 13
}
\keyword{datasets}
