% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample14.R
\docType{data}
\name{moodleRexample14}
\alias{moodleRexample14}
\title{Info for moodlequizR example 14}
\format{
A list
\describe{
  \item{quizname}{example14}
  \item{category}{MoodlequizR Examples / 14}
  ...
}
}
\usage{
moodleRexample14
}
\description{
A dataset containing the info to create the xml file for example 14
}
\keyword{datasets}
