% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample15.R
\docType{data}
\name{moodleRexample15}
\alias{moodleRexample15}
\title{Info for moodlequizR example 15}
\format{
A list
\describe{
  \item{quizname}{example12}
  \item{category}{MoodlequizR Examples / 15}
  ...
}
}
\usage{
moodleRexample15
}
\description{
A dataset containing the info to create the xml file for example 15
}
\keyword{datasets}
