% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample2.R
\docType{data}
\name{moodleRexample2}
\alias{moodleRexample2}
\title{Info for moodleR example 2}
\format{
A list
\describe{
  \item{quizname}{example2}
  \item{category}{MoodleR Examples / 2}
  ...
}
}
\usage{
moodleRexample2
}
\description{
A dataset containing the info to create the xml file for example 2
}
\keyword{datasets}
