% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample3.R
\docType{data}
\name{moodleRexample3}
\alias{moodleRexample3}
\title{Info for moodleR example 3}
\format{
A list
\describe{
  \item{quizname}{example3}
  \item{category}{MoodleR Examples / 3}
  ...
}
}
\usage{
moodleRexample3
}
\description{
A dataset containing the info to create the xml file for example 3
}
\keyword{datasets}
