% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample4.R
\docType{data}
\name{moodleRexample4}
\alias{moodleRexample4}
\title{Info for moodleR example 4}
\format{
A list
\describe{
  \item{quizname}{example4}
  \item{category}{MoodleR Examples / 4}
  ...
}
}
\usage{
moodleRexample4
}
\description{
A dataset containing the info to create the xml file for example 4
}
\keyword{datasets}
