% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample5.R
\docType{data}
\name{moodleRexample5}
\alias{moodleRexample5}
\title{Info for moodleR example 5}
\format{
A list
\describe{
  \item{quizname}{example5}
  \item{category}{MoodleR Examples / 5}
  ...
}
}
\usage{
moodleRexample5
}
\description{
A dataset containing the info to create the xml file for example 5
}
\keyword{datasets}
