% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample6.R
\docType{data}
\name{moodleRexample6}
\alias{moodleRexample6}
\title{Info for moodleR example 6}
\format{
A list
\describe{
  \item{quizname}{example6}
  \item{category}{MoodleR Examples / 6}
  ...
}
}
\usage{
moodleRexample6
}
\description{
A dataset containing the info to create the xml file for example 6
}
\keyword{datasets}
