% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample7.R
\docType{data}
\name{moodleRexample7}
\alias{moodleRexample7}
\title{Info for moodleR example 7}
\format{
A list
\describe{
  \item{quizname}{example7}
  \item{category}{MoodleR Examples / 7}
  ...
}
}
\usage{
moodleRexample7
}
\description{
A dataset containing the info to create the xml file for example 7
}
\keyword{datasets}
