% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample8.R
\docType{data}
\name{moodleRexample8}
\alias{moodleRexample8}
\title{Info for moodleR example 8}
\format{
A list
\describe{
  \item{quizname}{example8}
  \item{category}{MoodleR Examples / 8}
  ...
}
}
\usage{
moodleRexample8
}
\description{
A dataset containing the info to create the xml file for example 8
}
\keyword{datasets}
