% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.R
\name{nm}
\alias{nm}
\title{nm}
\usage{
nm(x, w, eps, ndigits, pts = 1)
}
\arguments{
\item{x}{vector of values}

\item{w}{list of weights}

\item{eps}{vector of precision}

\item{ndigits}{answers have to be rounded to ndigits, otherwise gives partial credit. Overrides eps}

\item{pts}{how many points is question worth?}
}
\value{
a character vector with the code for a CLOZE question
}
\description{
This function generates the code for a numerical CLOZE question
}
\examples{
nm(50)
nm(c(50, 40), w=c(100, 50))
}
