% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.data.R
\name{paste.data}
\alias{paste.data}
\title{paste.data}
\usage{
paste.data(sep = "", header = TRUE, is.table = FALSE)
}
\arguments{
\item{sep}{symbol used for separation}

\item{header}{does data have a header?}

\item{is.table}{is data a table? Needed if all data is character.}
}
\value{
the data in the clipboard
}
\description{
This function is used to read data from moodle into R
}
\examples{
# not run
# paste.data()
}
