% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png64.R
\name{png64}
\alias{png64}
\title{png64 Function}
\usage{
png64(plt)
}
\arguments{
\item{plt}{some graph object}
}
\value{
a character vector
}
\description{
This function creates a plot object that can be used in a moodle quiz
}
\examples{
library(ggplot2)
x=2*pi*1:1000/1000
df=data.frame(x=x,y=sin(x))
plt=ggplot(data=df,aes(x,y))+geom_line()
plt64 <- png64(plt)
}
