% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcategorical.R
\name{rcategorical}
\alias{rcategorical}
\title{rcategorical}
\usage{
rcategorical(n, p)
}
\arguments{
\item{n}{sample size}

\item{p}{vector or matrix of values}
}
\value{
a vector or a matrix
}
\description{
This function generates data from a univariate or a bivariate discrete distribution
}
\examples{
p=1:3
names(p)=letters[1:3]
x=rcategorical(1000, p)
p=matrix(1:6, 2, 3)
dimnames(p)=list(c("A","B"), letters[1:3])
x=rcategorical(1000, p)
}
