% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RtoHTML.R
\name{RtoHTML}
\alias{RtoHTML}
\title{RtoHTML}
\usage{
RtoHTML(method, x, y, n, varnames, ...)
}
\arguments{
\item{method}{name of the R routine}

\item{x}{data passed to all functions}

\item{y}{data passed to functions t.test (two-sample) and lm}

\item{n}{data passed to function binom.test}

\item{varnames}{names of variables as they are shown in quiz}

\item{...}{additional arguments passed to method}
}
\value{
a string
}
\description{
This function creates the code needed to make the output of selected R function appear correctly in moodle quizzes.
}
