% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa.R
\name{sa}
\alias{sa}
\title{sa}
\usage{
sa(txt, w = 100, caps = TRUE, pts = 1)
}
\arguments{
\item{txt}{character vector with possible answers}

\item{w}{vector of weights}

\item{caps}{keep capital letters}

\item{pts}{points for answers}
}
\value{
a character vector
}
\description{
This function creates a text question for moodle in CLOZE format.
}
\examples{
sa("Los Angeles")
sa(c("Los Angeles", "San Francisco"), w=c(100, 80))
}
