% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imoodleR.R
\name{shinymoodlequizR}
\alias{shinymoodlequizR}
\title{shinymoodlequizR}
\usage{
shinymoodlequizR()
}
\value{
None
}
\description{
This function runs the moodlequizR shiny app
}
\examples{
# not run
# shinymoodlequizR()
}
