% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genquiz.R
\name{genquiz}
\alias{genquiz}
\title{genquiz}
\usage{
genquiz(k = 1, fun, folder, problem = 0, funname, Show = FALSE, ...)
}
\arguments{
\item{k}{=1, how many quizzes?}

\item{fun}{name of the R routine that makes a quiz}

\item{folder}{where is the .R located?}

\item{problem}{(optional) which problem should be done?}

\item{funname}{name of quiz}

\item{Show}{=FALSE (optional) want to see what it looks like?}

\item{...}{further arguments passed to fun}
}
\value{
None
}
\description{
This function generates an xml file for import into moodle.
}
