% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.xml.R
\name{make.xml}
\alias{make.xml}
\title{make.xml}
\usage{
make.xml(fun, k = 1, folder, ...)
}
\arguments{
\item{fun}{(unquoted) name of function that makes a quiz, or number of a quiz}

\item{k}{how many quizzes?}

\item{folder}{folder were fun.R is located}

\item{...}{further arguments passed to fun}
}
\value{
None
}
\description{
This function is a simple wrapper for genquiz. It reads file from folder and runs genquiz. The default is to then remove the quiz.
}
