% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample1.R
\docType{data}
\name{moodleRexample1}
\alias{moodleRexample1}
\title{Info for moodleR example 1}
\format{
A list
\describe{
  \item{quizname}{example1}
  \item{category}{MoodleR Examples / 1}
  ...
}
}
\usage{
moodleRexample1
}
\description{
A dataset containing the info to create the xml file for example 1
}
\keyword{datasets}
