% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleRexample9.R
\docType{data}
\name{moodleRexample9}
\alias{moodleRexample9}
\title{Info for moodlequizR example 9}
\format{
A list
\describe{
  \item{quizname}{example9}
  \item{category}{MoodlequizR Examples / 9}
  ...
}
}
\usage{
moodleRexample9
}
\description{
A dataset containing the info to create the xml file for example 9
}
\keyword{datasets}
